requirements.txt
# =========================================
# AI Image Screener - Requirements
# Python 3.11+
# =========================================

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4

# File Type Detection
python-magic==0.4.27

# PDF Generation
reportlab==4.0.7

# ASGI Server Production
gunicorn==21.2.0

# Logging & Monitoring
colorama==0.4.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS & Middleware
starlette==0.27.0

# Testing (optional but recommended)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality (optional)
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Development Tools (optional)
ipython==8.18.1
ipdb==0.13.13

# =========================================
# Platform-Specific Notes:
# =========================================
# 
# Linux (Ubuntu/Debian):
#   sudo apt-get install -y libmagic1
#
# macOS:
#   brew install libmagic
#
# Windows:
#   pip install python-magic-bin==0.4.14
#   (alternative to python-magic for Windows)
#
# =========================================